test: consolidate the two E2E projects into WopiHost.E2ETests
The Collabora and ONLYOFFICE e2e projects were near-duplicates: DockerCheck and
PlaywrightFixture were byte-for-byte copies, and the two app fixtures shared the
whole boot-wait-probe-capture skeleton. One project now carries both suites:

- Fixtures/WopiAppFixtureBase absorbs the shared scaffolding (AppHost boot via
  the configureBuilder seam, resource waits, client readiness poll against the
  Aspire-discovered endpoint, RunProcessAsync/docker diagnostics, repo-root
  resolution). Derived fixtures supply only lane specifics: flags, resource
  names, readiness semantics (Collabora: /hosting/discovery 200; ONLYOFFICE:
  /healthcheck body "true"), timeouts, and client-log capture.
- Suites live under Collabora/ and OnlyOffice/, each tagged with a Client trait
  on top of Category=E2E. The nightly workflows run the same project with
  --filter "Client=Collabora" / "Client=OnlyOffice" (the runsettings bypass
  stays; once cleared, the CLI filter applies alone).
- Both collections keep DisableParallelization=true, so an unfiltered local run
  boots the two Aspire stacks sequentially, never concurrently (they pin the
  same host ports).

The repo's "duplicate PlaywrightFixture deliberately" precedent (SmokeTests vs
e2e) doesn't apply between the two e2e suites — they have identical dependency
surfaces — so merging removes real duplication without coupling anything new.
Also halves the per-PR compile cost (one Aspire.Hosting.Testing project in
WOPI.slnx instead of two).

Verified: all three tests pass locally through the exact workflow commands
(Client=Collabora: open + save; Client=OnlyOffice: open).
