+++ .runsettings
@@ -4,15 +4,15 @@
   Directory.Build.props; dotnet test does NOT auto-discover .runsettings by convention.
 
   TestCaseFilter excludes tests carrying the "E2E" trait so plain `dotnet test` skips the
-  Collabora end-to-end suite by default. That suite needs Docker + a real Collabora container
-  cold-start + Playwright; running it on every PR would burn CI minutes and gate PRs on a
-  flaky integration whose failures are mostly upstream.
+  end-to-end suites (Collabora + ONLYOFFICE) by default. Those suites need Docker + a real
+  client-container cold-start + Playwright; running them on every PR would burn CI minutes
+  and gate PRs on flaky integrations whose failures are mostly upstream.
 
   To opt back in for a single run, pass an MSBuild property that clears
   RunSettingsFilePath, which bypasses this file. Be aware that a CLI test filter does NOT
   work as an override here, because dotnet test ANDs the CLI filter with the runsettings
   filter (yielding (Category!=E2E)&(Category=E2E), which matches zero tests).
-  See test/WopiHost.E2ETests.Collabora/README.md for the full local-run command.
+  See test/WopiHost.E2ETests/README.md for the full local-run command.
 -->
 <RunSettings>
   <RunConfiguration>
