+++ Directory.Build.props
@@ -20,8 +20,8 @@
 		<GenerateDocumentationFile>true</GenerateDocumentationFile>
 		<!-- Auto-load the repo-root .runsettings for `dotnet test`. dotnet test does NOT
 		     auto-discover .runsettings by convention — this property is the documented seam.
-		     The runsettings currently carries a TestCaseFilter that excludes the E2E suite
-		     from default runs; see .runsettings and test/WopiHost.E2ETests.Collabora/README.md. -->
+		     The runsettings currently carries a TestCaseFilter that excludes the E2E suites
+		     from default runs; see .runsettings and test/WopiHost.E2ETests/README.md. -->
 		<RunSettingsFilePath>$(MSBuildThisFileDirectory).runsettings</RunSettingsFilePath>
 	</PropertyGroup>
