--- test/WopiHost.E2ETests.OnlyOffice/WopiHost.E2ETests.OnlyOffice.csproj
@@ -1,39 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFramework>net10.0</TargetFramework>
-		<IsPackable>false</IsPackable>
-		<IsTestProject>true</IsTestProject>
-		<!-- Both the Web frontend and the Validator carry wwwroot/file.ico; the static-asset
-		     aggregator has no way to pick a winner when both are pulled in transitively via
-		     the AppHost reference. We don't serve static assets from the test runner anyway. -->
-		<StaticWebAssetsEnabled>false</StaticWebAssetsEnabled>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<!-- Reference the AppHost itself (not its project graph) so DistributedApplicationTestingBuilder
-		     can boot the whole resource topology from inside a test. IsAspireProjectResource="false"
-		     means: I want the AppHost's API surface (Projects.WopiHost_AppHost generated type), but
-		     don't treat it as an Aspire resource — the test IS the host. -->
-		<ProjectReference Include="..\..\infra\WopiHost.AppHost\WopiHost.AppHost.csproj" IsAspireProjectResource="false" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Aspire.Hosting.Testing" />
-		<PackageReference Include="Microsoft.Playwright" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Update="coverlet.collector">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
-		<PackageReference Update="coverlet.msbuild">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
-		<PackageReference Update="xunit.runner.visualstudio">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
-	</ItemGroup>
-</Project>
