+++ test/WopiHost.E2ETests/Fixtures/DockerCheck.cs
@@ -1,11 +1,11 @@
 using System.Diagnostics;
 
-namespace WopiHost.E2ETests.OnlyOffice.Fixtures;
+namespace WopiHost.E2ETests.Fixtures;
 
 /// <summary>
-/// Probe for whether a usable Docker engine is reachable from this process. The ONLYOFFICE e2e
-/// fixture short-circuits with a skip when this returns <see langword="false"/> — contributor
-/// machines without Docker Desktop running shouldn't see test failures, just a clean skip.
+/// Probe for whether a usable Docker engine is reachable from this process. The e2e app fixtures
+/// short-circuit with a skip when this returns <see langword="false"/> — contributor machines
+/// without Docker Desktop running shouldn't see test failures, just a clean skip.
 /// </summary>
 internal static class DockerCheck
 {
@@ -14,6 +14,12 @@ internal static class DockerCheck
     /// non-empty server version as "Docker is available". A missing CLI, a Windows-only client
     /// without a running engine, or a context pointing at an offline daemon all surface here.
     /// </summary>
+    /// <remarks>
+    /// The check is intentionally minimal — it only answers "can Aspire start containers right
+    /// now". It does <em>not</em> verify Linux-container mode, Docker Compose, image pull
+    /// permissions, or any other capability. A more specific path failing downstream should
+    /// surface the real error rather than be masked by a coarser preflight.
+    /// </remarks>
     public static bool IsDockerAvailable()
     {
         try
