+++ test/WopiHost.E2ETests/OnlyOffice/OnlyOfficeFixtureCollection.cs
@@ -1,4 +1,4 @@
-using WopiHost.E2ETests.OnlyOffice.Fixtures;
+using WopiHost.E2ETests.Fixtures;
 
 namespace WopiHost.E2ETests.OnlyOffice;
 
@@ -7,11 +7,14 @@ namespace WopiHost.E2ETests.OnlyOffice;
 /// (<see cref="OnlyOfficeAppFixture"/>) with the Playwright browser
 /// (<see cref="PlaywrightFixture"/>). Members of this collection share a single ONLYOFFICE
 /// container instance — that's where the cold-start cost is amortised.
+/// <c>DisableParallelization</c> also keeps this collection from running concurrently with the
+/// Collabora one: each boots its own Aspire stack on the same pinned host ports.
 /// </summary>
 [CollectionDefinition(Name, DisableParallelization = true)]
 public sealed class OnlyOfficeFixtureCollection
     : ICollectionFixture<OnlyOfficeAppFixture>,
       ICollectionFixture<PlaywrightFixture>
 {
+    /// <summary>Collection name referenced by the suite's test classes.</summary>
     public const string Name = nameof(OnlyOfficeFixtureCollection);
 }
